Moved test helpers
