+++ components/bpmn-q/test/helpers/ModelerHelper.js
@@ -1,6 +1,6 @@
-import {createTempModeler} from '../modeler-component/editor/ModelerHandler';
-import {loadDiagram} from '../modeler-component/common/util/IoUtilities';
-import {setPluginConfig} from '../modeler-component/editor/plugin/PluginConfigHandler';
+import {createTempModeler} from '../../modeler-component/editor/ModelerHandler';
+import {loadDiagram} from '../../modeler-component/common/util/IoUtilities';
+import {setPluginConfig} from '../../modeler-component/editor/plugin/PluginConfigHandler';
 
 const app1 = {
   id: 'app1',
